frontend/requirements.txt
gradio>=4.20.0
requests>=2.31.0
Pillow>=10.0.0
